Merged PR 260485: Migrate the window-post-message-proxy library from TSLint to ESLint
Completed the task of migrating from TSLint to ESLint.

1. Removed the TSLint configurations and added configurations for ESLint.
2. Updated the default Linter to ESLint.

The base `.eslintrc.js` is taken from the [JS SDK](https://dev.azure.com/powerbi/Embedded/_git/powerbi-javascript?path=/.eslintrc.js).

The following lines have been changed:
1. ![image (6).png](https://dev.azure.com/powerbi/4c7b5adb-c2d0-4f18-b23c-edc4ac30f4e1/_apis/git/repositories/e063268f-1107-4b90-a0c0-1785e0769499/pullRequests/260485/attachments/image%20%286%29.png)
 for using the main `tsconfig` file instead of the test file.

2. ![image (7).png](https://dev.azure.com/powerbi/4c7b5adb-c2d0-4f18-b23c-edc4ac30f4e1/_apis/git/repositories/e063268f-1107-4b90-a0c0-1785e0769499/pullRequests/260485/attachments/image%20%287%29.png)
as discussed in the email, any rules that needs code change need to be ignored.

3. ![image (8).png](https://dev.azure.com/powerbi/4c7b5adb-c2d0-4f18-b23c-edc4ac30f4e1/_apis/git/repositories/e063268f-1107-4b90-a0c0-1785e0769499/pullRequests/260485/attachments/image%20%288%29.png)
as discussed in the email, any rules that needs code change need to be ignored.

4. ![image (9).png](https://dev.azure.com/powerbi/4c7b5adb-c2d0-4f18-b23c-edc4ac30f4e1/_apis/git/repositories/e063268f-1107-4b90-a0c0-1785e0769499/pullRequests/260485/attachments/image%20%289%29.png)
Rule in the `tslint.json`

5. ![image (10).png](https://dev.azure.com/powerbi/4c7b5adb-c2d0-4f18-b23c-edc4ac30f4e1/_apis/git/repositories/e063268f-1107-4b90-a0c0-1785e0769499/pullRequests/260485/attachments/image%20%2810%29.png)
Rule in the `tslint.json`

There are no warnings or errors found in `src` folder the console after running ESLint as seen in the following image:

![image (2).png](https://dev.azure.com/powerbi/4c7b5adb-c2d0-4f18-b23c-edc4ac30f4e1/_apis/git/repositories/e063268f-1107-4b90-a0c0-1785e0769499/pullRequests/260485/attachments/image%20%282%29.png)

There are no warnings or errors found in `test` folder the console after running ESLint as seen in the following image:

![image (4).png](https://dev.azure.com/powerbi/4c7b5adb-c2d0-4f18-b23c-edc4ac30f4e1/_apis/git/repositories/e063268f-1107-4b90-a0c0-1785e0769499/pullRequests/260485/attachments/image%20%284%29.png)

After running npm audit, we get 5 vulnerabilities as shown in the picture below. All the vulnerabilities are caused by using the typings library.

![image (3).png](https://dev.azure.com/powerbi/4c7b5adb-c2d0-4f18-b23c-edc4ac30f4e1/_apis/git/repositories/e063268f-1107-4b90-a0c0-1785e0769499/pullRequests/260485/attachments/image%20%283%29.png)

Related work items: #792512, #795799