Merged PR 570838: [Fix S230 alerts] - removed gh-pages and replaced gulp-header with gulp-prepend
removed gh-pages and replaced gulp-header with gulp-prepend, both depend on lodash.templae which has a vulnerability in all of its version.

gh-pages is not required anymore as we post public docs to MS docs and not github.
replaced gulp-header with gulp-prepend which serves the same purpose - adds the header to generated d.ts files. Example of the header:
/*! window-post-message-proxy v0.2.8 | (c) 2016 Microsoft Corporation MIT */

Related work items: #1379776